Improve SawpErrorState method/trait
`SwapErrorState` has two methods and a trait.  Since this object stores
boolean values the methods to set and get elements should be named in a
manner that makes it explicit what the boolean value means.  The getter
method is functionally a predicate function so naming it as such makes
the code easier to read.

We currently have a trait that calls through to the getter method, we
can simplify this by just having the trait and implementing the trait
instead of the getter method.

At call sites for the getter method a local variable is named
`swap_has_failed`, this hints at the meaning of this getter.  Re-name
the getter trait to `has_failed` to match this usage.