Make the code more braindead
There is no obvious reason why the `swap` local variable is hidden in
the block that sets tuple `(alpha_htlc_params, beta_htlc_params)`.

Using a tuple that is set by a block that returns it makes developers
have to look twice because it is not 'typical'.

If we make the code more braindead simple it is easier for future
developers to read, said another way, it requires less cognitive load to
read the code - this is a good thing.