Add check to see if HTTP port is available
If the HTTP port is in use we would like to exit cnd.  We can check if
the port is available before trying to spawn the HTTP server.  This
patch introduces a race condition between the call to
`port_is_available` and the actual spawn of the wrap server.  If we hit
this race condition it is non-critical, it just means cnd will start
without an HTTP server, the logs will still contain the bind error.